utils.cpp/hpp cleanup
